Do not reformat system messages
When the user sends logging messages using `add_log` and associated
functions, he has full control over how it will be formatted via
`format_message`. This is not the case for intercepted system messages
that were sent with `stop`, `warning`, and `message`, so
`format_message` should never try to to reformat them.